// src/generated/extensions-v2/types.gen.ts
export type ExtensionListResponse = {
  result: Array<ExtensionListItem>;
  pagination: {
    next_cursor: string | null;
    has_more: boolean;
  };
};

export type ExtensionListItem = {
  id: string;
  type:
    | 'mod'
    | 'theme'
    | 'payment-gateway'
    | 'server-manager'
    | 'domain-registrar'
    | 'hook'
    | 'translation';
  name: string;
  description: string;
  website: string;
  license: License;
  icon_url?: string;
  source: Repository;
  version: string;
  download_url: string;
  developer: PublicDeveloper;
};

export type License = {
  name: string;
  URL?: string;
};

export type Repository = {
  type: 'github' | 'gitlab' | 'custom';
  repo: string;
};

export type PublicDeveloper = {
  id: string;
  type: 'user' | 'organization';
  name: string;
  URL?: string;
  avatar_url?: string;
  approved: boolean;
};

export type Error = {
  error: {
    message: string;
    code: string;
    details?: Array<unknown>;
  };
};

export type Extension = ExtensionPayload & {
  developer: PublicDeveloper;
};

export type Release = {
  tag: string;
  date: string;
  download_url: string;
  changelog_url?: string;
  min_fossbilling_version: string;
};

export type ExtensionPayload = {
  id: string;
  type:
    | 'mod'
    | 'theme'
    | 'payment-gateway'
    | 'server-manager'
    | 'domain-registrar'
    | 'hook'
    | 'translation';
  name: string;
  description: string;
  releases: Array<Release>;
  website: string;
  license: License;
  icon_url?: string;
  readme: string;
  source: Repository;
  version: string;
  download_url: string;
};
// ...
export type GetExtensionsData = {
  body?: never;
  path?: never;
  query?: {
    type?:
      | 'mod'
      | 'theme'
      | 'payment-gateway'
      | 'server-manager'
      | 'domain-registrar'
      | 'hook'
      | 'translation';
    developer_id?: string;
    limit?: number;
    /**
     * Opaque cursor returned by the previous page
     */
    cursor?: string;
  };
  url: '/extensions';
};